crawler/lanuv/requirements.txt
pandas
psycopg2_binary
Shapely
sqlalchemy
